Refactor spatial objective functions into separate module
- Moved spatial objective creation functions from parse_model_lucent.py to a new file spatial_objectives.py for better code organization and maintainability.
- Updated import statements in parse_model_lucent.py to utilize the new spatial_objectives module.
- Removed redundant code and ensured that the functionality remains intact.
